requirements.txt
# Bookscart - core dependencies (pinned to tested-ish versions as of Sep 2025)
Django==5.2.6
djangorestframework==3.16.1

# Database / drivers
psycopg-binary==3.2.9        # for Postgres (production/demo)
duckdb==1.3.2                # DuckDB / MotherDuck client

# Media / Cloudinary
cloudinary==1.44.0
django-cloudinary-storage==0.3.0

# Authentication
django-allauth==65.11.1

# API / async / realtime
djangorestframework-simplejwt==5.5.1
channels==4.3.1
channels-redis==4.3.0
uvicorn==0.35.0
gunicorn==23.0.0
whitenoise==6.9.0
pillow==11.3.0
django-cors-headers==4.7.0

# AI / Agent / Vector DB
langchain==0.3.27
langgraph==0.6.6
openai==1.106.1
chromadb==1.0.20
chromadb-client==1.0.20

# HTTP / utilities
httpx==0.28.1
requests==2.32.5
python-dotenv==1.1.1
django-environ==0.12.0

# Optional integration helpers (use if needed)
langchain-openai==0.3.32     # optional LangChain <-> OpenAI helpers
langchain-community==0.3.29
cryptography==45.0.7


